Introducing v4
1) moved stepper motor code from subsystem_main.ino to core_stepper_motor.ino via static variables and a setup() function; successful
2) tried to implement Gabe's improved ultrasonic_sensor.ino into the folder with car_main.ino. Had to write a setUpUltrasonicRangefinder() function, and then changed line 115 from 'unsigned int dist_mm' to 'dist_mm', since the variable had already been initialized earlier. The new function leads to erratic behavior of the robot.

Description of erratic behavior:
In line 78 of car_main.ino (v4 folder), the code checks the distance in front of the ping sensor to see if it is less than 200mm. If it is less than this, the robot believes there is a nearby obstacle, and takes action.

Lines 57 - 61 of ultrasonic_sensor.ino initialize the raw distance and the filtered distance to be zero, which is less than the value checked for in car_main.ino. This causes the robot to take action as if there were something in front of the sensor, although there is nothing close by. After completing the function that is supposed to occur when something is close (stopping and then either 'shaking the head no' or 'doing a dance move'), the robot 'senses' something in front two more times before behaving correctly--aka, not taking action until there is something in front.

Apart from behaving erratically at the startup, the robot also take several seconds to sense an actual object in front of the sensor, as if waiting for the median of the 5 samples to trigger correctly.

See the serial plotter output from lines 112 to 114 of ultrasonic_sensor.ino.

Version 3 of the code behaves correctly--aka, does not have erratic behavior at the startup of the robot, and responds almost immediately when a hand or object is placed in front of the sensor.  Copying the ping sensor code from v3 would fix the problem, although I'm hoping to get the improved ping sensor code working, since it is supposed to have fewer erratic peaks and valleys.

Suspecting that the issue may lie in the delay() function from car_main.ino, an attempt at a fix was made in the way of an improved delay function that samples ultrasonic sensor code in the background. This function is entitled 'newDelay()'; however, when implemented, instead of just having 3 erratic triggers of the robot's sense of an object being close, it made it trigger the robot at least 7 times at the beginning. It also caused erratic trips of the proximity code every 20 seconds or so, with no regularity that I could see.

It would probably be good to have the proximity sensor initialize at some value other than zero; probably initialize it to just higher than the value it determines to be 'nothing out in front'; but, it still wouldn't fix the problem of not detecting an actual object in front of the sensor quickly.

One more item of note--it is quite humid here at scout camp where I'm making these changes. I hope the humidity isn't playing a factor with the robot hardware.