Synced with new computer files
I realized that the files on my new computer weren't connected to GitHub. Loaded a v0_5_0 to the generic robot code (although I'm not sure there were any changes from v0_4_0). Main changes are to a file called rubber_band_bot (started my own folder with file specific to a rubber band bot, not the generic file that can be used for any type of robot).